feat(ui): implement ViewModels and update Entities

+++ app/src/main/java/com/pwlimaverde/orcafacilapp/data/local/entity/BudgetItem.kt
@@ -32,5 +32,5 @@ data class BudgetItem(
     val valorUnitario: Double,
 
     @ColumnInfo(name = "quantidade")
-    val quantidade: Int
+    val quantidade: Double
 )
